reflekt-registry/requirements.txt
ansicon==1.89.0 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
attrs==21.4.0 ; python_version >= "3.9" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
blessed==1.20.0 ; python_version >= "3.9" and python_version < "4.0"
botocore==1.29.87 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
chalice==1.27.3 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
defusedxml==0.7.1 ; python_version >= "3.9" and python_version < "4.0"
httpie==3.2.1 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
inquirer==2.10.1 ; python_version >= "3.9" and python_version < "4.0"
jinxed==1.2.0 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.9" and python_version < "4.0"
markdown-it-py==2.2.0 ; python_version >= "3.9" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.9" and python_version < "4.0"
monotonic==1.6 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
pip==22.2.2 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.9" and python_version < "4"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-editor==1.0.4 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
readchar==4.0.3 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==0.10.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
requests[socks]==2.28.2 ; python_version >= "3.9" and python_version < "4"
rich==13.3.2 ; python_version >= "3.9" and python_version < "4.0"
segment-analytics-python==2.2.2 ; python_version >= "3.9" and python_version < "4.0"
setuptools==67.6.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.9" and python_version < "4.0"
wheel==0.38.4 ; python_version >= "3.9" and python_version < "4.0"
